Copy over package.json from cssbundline-rails tailwind installation for Intellisense
The Intellisense integration with Tailwind only works when it is installed vie npm, yarn or pnpm
https://github.com/tailwindlabs/tailwindcss-intellisense?tab=readme-ov-file#troubleshooting

However, the default Rails tailwind installation on `rails new` and also the “tailwindcss-rails” gem use a self-contained standalone tailwindcss executable

https://github.com/rails/tailwindcss-rails?tab=readme-ov-file#installation
https://github.com/flavorjones/tailwindcss-ruby

This version does not work with intellisense.

Although it may be a good idea to use the standalone version for deployment and also for building in development, not working with Intellisense is a pain. This allows you to have the best of both worlds.

You will have a node version for Intellisense, and the standalone version for all the rest.